Work on GMST renumbering; work on DOBJ support
The GMST renumber process now checks for any inbound references to the GMST and attempts to sever them. If they cannot be severed, an error code is returned. Additionally, there is now an error code for trying to renumber a GMST not defined in the active file.

When attempting to retrieve the canonical stub for a singleton form, you can now create a stub if none exists. This will help with editing.

Added a loaded form class for DOBJ.

Added a helper class to the form-load interface, to make it faster and easier to check whether the record being loaded is from the active file (if any).

Added some frontend functions for retrieving the name and description of a DOBJ entry. The internal list is extremely incomplete.