Work on DOBJ support
Dumped all DOBJ names and form types from the executable. Names and descriptions will be located in frontend code; signature-to-form-type mappings will be in the backend.

Fixed (file_load_order::get_canonical_instance_of_singleton_form). One of the overloads was reliant on the other, but due to a const-related ambiguity, wasn't calling it properly (i.e. it was calling itself instead).