Numerous fixes to hardcoded forms; more DOBJ work
The on-demand form load process was broken for hardcoded forms, and in several ways. For the below explanation, assume that every hardcoded form's stub has a loaded-form instance created for it during the file load process (as was originally intended).

Firstly, we would never load any overrides of hardcoded forms, because (form_stub::_load) would see that the stub already had a loaded-form pointer and would therefore stop early.

Secondly, if a hardcoded form was overridden, then the loaded-form instance created for it would be lost, because when one stub overrides another, we delete the overridden stub.

Thirdly: the loss of a hardcoded form's stub's loaded-form instance would allow it to attempt to load, but this attempt would fail, because the dummy file used for hardcoded form data can't actually load anything.

Fourth: if a hardcoded form was overridden, then it would be permitted to unload. However, even if the above issues were addressed individually, this would mean that any hardcoded loaded-form data would be lost. That data was only set up one time (when the loaded-form instance was first created).

All of the above issues have been fixed by making it so that hardcoded form data is instantiated on-demand. Additionally, this allows us to unload hardcoded forms.

Separately from all of the above, more work has been done on DOBJ support, including making the DefaultObjectManager flag itself as "edited" when setting any entry. The DOBJ UI also now listens for the signal sent when a DOBJ entry is modified.