Critical fix to singleton form loading
Fixed an edge-case that would occur when loading an override of a singleton form: the old, deleted form stub still had its pointer in the form maps, so it could be found and mistaken for the canonical form stub (a use-after-free error), corrupting the file list of the new form stub.

Added read/write methods for "signatures" within subrecord bodies which will handle byteswapping when needed; currently, DefaultObjectManager makes use of this.