Minor debug helper check
Added a debug check to (file_load_order::save_active_file): if any active file form stub has a non-zero refcount, then we'll break into the debugger if we're compiled in Debug.

We don't need this check -- we already take exactly the precautions it's meant to encourage -- but it will likely be useful should anyone else want to use our backend.