Fixes to post-save cleanup re: non-serializable forms
Post-save cleanup will now properly sever references to non-serializable forms that cannot be deleted. A good example of this case: imagine converting a Skyrim Special file to Skyrim Classic, where the Special file contains a FormList that refers to a VOLI defined in one of its masters. We can't delete the VOLI because it's not in the active file, so we need to sever references to it manually instead.

Post-save cleanup is now more precise about how it discards none-stubs: they will be discarded if they have no inbound references or if all of their inbound references meet one of two conditions: they are outbound from forms that are themselves going to be discarded; or they are outbound from forms that were serialized and thus are subject to the post-save cleanup change described above.