Optimized the post-sae none-stub discard
Optimized the post-save none-stub discard: we don't need to check all of a none-stub's inbound references against the fixup data, because the fixed-up forms would've had their outbound references to the none-stubs severed as part of the fixup step... which means that none of a none-stub's inbound references should ever be from any of those forms.

Additionally, before we scan over none-sutbs, we cache the current size of the (stubs_to_remove) list. Why? Because all stubs past that point will be none-stubs that we've added while looping, and none-stubs can't refer to other forms, so we only need to check the (stubs_to_remove) list up to that point.