Improvements for injected records and the object window
Added getters to (form_stub) to check whether it's an edited or active file form, and to check whether it's an injected form.

The Object Window now displays asterisks on the editor IDs of both edited and active file forms (the distinction is that if you load an active file and don't edit its forms, those forms now have asterisks). This is more consistent with the Creation Kit, which shows asterisks on all such forms' use counts.

The Object Window now queries most display-related properties of form stubs when acquiring or updating its pointers to them, instead of at display time.

The Object Window now displays the form IDs of injected records in green, using a brighter shade for active file forms and a darker shade for non-active-file forms.

When listing none-stubs, the Object Window now italicizes the "<non-existent form>" string, to further distinguish it from a real editor ID.