Record injection now double-checks none-stubs' "cause" files
DovahKit no longer allows you to inject records overtop none-stubs, if those none-stubs are the result of one or more dangling references inbound from forms outside of the active file.

Injected forms' IDs are always shown using a consistent shade of green now. We don't need to distinguish the color based on whether they're from the active file, because we already show an asterisk next to the name and form ID for that purpose.