Work on reorganizing the file load process
Implemented a system whereby subclasses of (file_threaded_part_loader_base) can specify recommended thread counts for their type, by way of a static class member; a templated constexpr int can then be used to query these for any given type. On reflection, this... may be overengineered. I'll strip it down in the next commit, probably.

Copied over the old code for threaded readers.

Since we no longer split threaded readers into different lists by type, the (file_loader) class now has a helper function for retrieving the Nth reader of any given type. The helper function is templated, allowing us to check the recommended thread count for the given reader and automatically wrap the index passed in; this makes it a bit easier (read: lazier) to maintain some of the counters we use when sorting pending work into each of the different threads. Having to iterate over the list of all threads is going to be slower than having separate lists, but not by much (we only have 18 threads with the current configuration), and it makes the code more maintainable.

Copied over the (file_reader) code for opening a mapped file and for running the file load process.

Currently, we have a bunch of errors stemming from the fact that functions for interfacing with the file load order -- mainly, creating stubs and storing them -- aren't yet ported over.