Work on reorganizing the file load process
Created the (form_stub_build_interface) for (file_loader); both it and its worker-thread objects use the interface to create stubs, commit them to the (file_load_order), and extract their high-value subrecords (that is, any that we want to grab for the form stub itself).

We're still not done. We're getting there.