New file-loading code now runs
Tested and made several fixes until we could load Skyrim.esm. It takes 1.2 times as long as the old code, so if I can't find out why that is, then I'm probably just going to roll entirely back to the old code.

I was never even really intending to remodel it all. I just wanted to improve how it reports errors, and I guess I just got... lost. It was (sort of unavoidably?) a bit tangled, and I thought it'd be easier to try and perfect it. Obviously not.