Slowdown resolved.
The slowdown has been resolved. There were a few places where we were blindly resetting a detailed notice that (basic_reader) uses to retain error information. Now, we just reset it only when we ever actually catch an error. (The reason we want to reset it at all is so that hitting, say, a bad 'XXXX' subrecord during one on-demand form load doesn't permanently break all on-demand form loads going forward.)

I'm still not even remotely *happy* with the file-loading code and the way it's organized, but we're back to it being as snappy and efficient as it was before, and now it shares a common system for reporting errors and warnings. Really, the reason I'm unhappy with it is because it's basically the same as it was before. I've split a few classes up, but if I were to diagram the old system and lay that on top of the new system's diagram, it'd be the exact same shape, with the exact same unorthodox connections and self-pointing objects.