Started work on moving record flags to form stub source file info
Started work on moving record flags from the loaded form classes to the form stubs' source file list entries. We need this in order to be able to tell whether a form is defined as a partial record in any given file.

To that end, the (form_stub) class now has a setter for these flags, which will flag the form as edited, add the active file to its file list (if needed) with an offset of 0, and set the flags.