Fixed a crash in (file_loader::close)
Fixed a crash in (file_loader::close) stemming from it blindly calling (file_threaded_part_loader_base::wait_for), which in turn called (std::thread::join) without first checking whether the thread is joinable (outside of a load attempt, it wouldn't be). The crash occurred when saving, as we need to close the file in order to replace it with newly-saved data.