Resuming work on the form save interface
More form components now take the form save interface when saving. None of them make any use of it yet, but they could depending on what we have to add in the future, and in any case I'd prefer for form components to be called and used consistently with one another.