Editor scripting: work on cross-thread messaging
Fixed a crash-on-exit that occurred due to leaving a std::thread unjoined.

The script VM now has a basic system for providing APIs to the userscript.

Tests show that we're properly passing messages, or so it seems; the log widget only shows numbers (which almost look like items in a vertical header?) instead of the text we try to log.