Editor scripting: confirmed to work: one-way messaging; force-kill
Fixed the issues with the script window's log: it's a QTableWidget, so its column count needed to be set to 1 in Qt Designer (the default was 0). With this fix made, the (dovah.log_message) Lua API now properly logs messages to the UI.

Added some basic style code for the script window's log pane.

Fixed another crash-on-exit, again from std::thread joining shenanigans. Now it should be good.

Fixed a mistake that caused us to accidentally set (DovahKitScriptVM::running) to false immediately after setting it to true when running a script. We need to not set it to false in (DvoahKitScriptVM::_teardown_lua_vm). With this fix made, we can now force-kill a script from the main thread, with proper cleanup, as designed.