Editor scripting: script-to-main messages can now block the script thread
The script thread is now capable of sending a message to the main thread and waiting until the message is acknowledged.

The script window now has a title.

Cleaned up the code that manages the main thread tick handler, and the code that handles sending messages from Lua APIs to the main thread.