Editor scripting: Work on cross-thread messaging
Reorganized the script VM's message queues; it is now obvious from their names what each queue's sender and recipient is. We now prefer "script-to-main" and "main-to-script" over "outbound" and "inbound," which were ambiguous.

Separated the VM's script-thread loop into a member function, and wrote placeholder code to process the main-to-script message queues. This includes code to spin the script thread before termination if there's anything that should keep a script "alive," i.e. script-spawned UI (which isn't implemented yet).