Editor scripting: started work on userdata classes
Copied some helper functions for Lua classes over from an unfinished Skyrim project.

Added a shimmed version of the "collectgarbage" function, to allow scripts to demand garbage collection but not to tamper with how it's done.

Wrote up some plans for how we'll implement form and UI wrappers while taking advantage of the __gc metamethod. Wrote some skeletal class definitions for a framework for this.