Editor scripting: minor cleanup for wrappers; form.get_user_forms
Created a helper function which: checks whether a newly-created wrapper is redundant with an existing wrapper, and if so, replaces it with the existing one; and adds the wrapper to the Lua stack.

Optimized equality comparisons for wrappers.

Added a new API: (form.get_user_forms). It returns an array of all forms that use the callee form.