Editor scripting: wrapper cleanup
Cleaned up the implementation for Lua wrappers.

Lua userdata no longer consists of pointers to structs that we've allocated ourselves, but rather of the structs themselves; these userdata values are stored in a weak-table, making them accessible to native code. This means that we can now reuse the userdata itself when reusing wrappers, which in turn confers a number of benefits: a) we're no longer heap-allocating tons of small pointers to the same objects; b) when we need to "kill" a userdata, we can clear its metatable; c) we no longer need to track refcounts or otherwise do any memory management work for wrappers. Equally, there are costs: every time we return a new wrapper, we have to scan all extant wrappers in order to see if there already exists a wrapper for the same data.