Editor scripting: redesigned how wrappers work
We will no longer be subclassing (editor_script::wrapper). Instead, the Lua APIs themselves will access the wrapped data using private helper functions (i.e. non-member functions defined in a nameless namespace in the CPP files). The main "wrapper" class now has all of the information that a Lua API would need in order to actually access the wrapped data.

As part of the above change, metatable definitions have been decoupled from the wrapper class. There are now empty struct definitions (subclasses of (wrapper_metatable)) which cannot be instantiated and exist solely to allow easy templating of metatable-related code.

Defined a new type: cobb::eight_cc. It was created in response to a limitation in MSVC's multicharacter literal support, where 'ABCD' is valid but 'ABCDEFGH' is a syntax error even when stored in a uint64_t.