Editor scripting: finished wrapper redesign; form type support; fixes
Fixed a major oversight during the last userdata revision: we were still manually heap-allocating wrappers and rigging the userdata up to point to them, but since we were no longer doing manual memory management, that just meant we were leaking wrappers. Now, the wrappers actually *are* the userdata and so are managed by Lua automatically, as intended.

The script VM is now capable of returning forms to Lua using different metatables for different form types.

Wrappers now contain a loaded-form pointer. Forms are not loaded when the wrapper is created; rather, a load must be requested by each Lua API.

Wrappers now have a __gc metamethod, which is needed to ensure that the wrapper destructor is called (since Lua otherwise doesn't know or care what data is stored in userdata when it deletes that userdata).

Our Lua class boilerplate now forwards the __gc metamethod, as it must be forwarded to work on userdata.

Wrote a Lua class for voicetypes.