Editor scripting: high-precision timing API; for_each_form_of_type
Added Lua APIs for high-precision timing and for iterating over all forms of a given type.

Wrote a (cast_to_exact_class) function for Lua-to-C code. The high-precision timing API makes use of it in order to validate argument types as quickly as possible.

We don't currently load the "os" library, but I've added a member function whitelist for it, for if we ever do in the future.