Editor scripting: fixed class getters; performance tests on getters look good
Fixed the implementation of class getters.

The (define_wrapper_metatable) helper now uses any getter/setter lists on the class.

Added an (is_female) field to the voicetype class for testing. Benchmarks indicate that it tends to be 600 microseconds faster than the member function. I suspect that's because for both the field access and the member function call, there's a script command to look up a function on the metatable followed by that function being called, but in the former case, the function call happens in native code and in the latter case, it happens as a script command. Or something.