Editor scripting: wrapper::is_equal fixes; Lua utility class plans
Native code now prefers lua_isnoneornil over lua_isnil.

Wrappers that are flagged as being collections should now properly compare their last part.

Wrote incomplete code for a "vector3" Lua class, and wrote up a list of other utility classes that would be useful to offer to Lua (and that could be offered in native code as well). The goal behind the "vector3" class is that a wrapper metatable (e.g. for REFR positions) could subclass it and provide its own getters/setters for the "x", "y", and "z" fields, and the class methods on the "vector3" metatable would be able to access those (since they use lua_gettable and not lua_rawget).