Editor scripting: work on utility classes for scripts
Created a helper function for classes: check_for_class. It's similar to cast_to_class, except that the latter can only be used for userdata, whereas the former can be used for normal scripted tables that are backed by native classes, such as the vector3 utility class.

Implemented script APIs: (vector3.is), (vector3.new), (vector2.is), and (vector2.new).

Added __tostring metamethods to the vector2 and vector3 utility classes.