Editor scripting: completed __pairs for classes
The __pairs metamethod used for classes is now able to access inherited members properly.

We now have debugging functions for internal use; they can dump the contents of the Lua stack, including the full contents of nested tables. Tables are tagged with a number; tables that occur multiple times in the output show only the tag, both to avoid infinite recursion and to make the log smaller. These functions use lua_next, not any equivalent to pairs(), so if used on one of our userdata (or its metatable) they won't show members the way a Lua script would "see" them; this means that these functions aren't a suitable basis for a function that would imitate ESO's "d" function.