Editor scripting: Plans for wrapped collections
Now that we've had some practice implementing stateful iterators for __pairs, we have all the knowledge we need to plan out a sane implementation of wrapped collections (e.g. a list of a form's Papyrus scripts; a list of a form's destruction stages; etc.).

In preparation for this, (editor_script::wrapper::part) has been modified: the (index) and (name) fields are no longer a union. This means that for collections of named elements, you can access elements by name or by index. It also means that we can use std::string for the name, saving us the trouble of managing memory for those strings manually.