Editor scripting: Plans
Removed completed tasks from the to-do list.

Next task is handling removal from sequential collections, so that script variables that refer to collection elements by index don't fall out of synch. After that, we should be free to just start exposing as much form data to scripts as we can.