Editor scripting: #collection now works; wrappers no longer retain names
Collections now have a __len metamethod, and so now support the # operator.

Wrappers are no longer able to store string keys to collections. This simplifies the code needed to update sequential collection elements' indices when an element is deleted from the collection; it also simplifies a lot of wrapper-related code as a whole.

The main issue that the removal of names helps avoid is mismatches when actually deleting an element. From inside of the Lua API for removing a collection item (given the item's name, index, or an existing wrapper), we don't have a good way to check if a wrapper already exists for that item. The best we can do is create a new wrapper with matching data, and have a function in the VM that searches for a wrapper that is an exact match. The thing is, we can't know whether the existing wrapper (if any) uses a name or an index, and the VM core doesn't know how to convert between the two (because it doesn't actually understand the wrapped data). Moreover, there's a similar problem with potentially having multiple active wrappers for the same piece of underlying data, if it was accessed once by name and then once by index, and this problem has a similar cause. By forcing wrapper creation to always resolve names to indices and only allowing wrappers to retain indices and not names, we avoid this entire set of problems.