Editor scripting: Plans
The Lua API for removing Papyrus scripts from a form has to do a lot of work to resolve to a usable collection index to remove. Can we move much of that into a helper function?

Our code for "killing" a wrapper when the wrapped object is deleted is incomplete.