Editor scripting: simplified the code to remove a Papyrus script
Simplified the API for removing a Papyrus script. When the function is given an argument that isn't a wrapped script, it will access (self.scripts[arg]) instead of trying to extract an index from the argument itself. This makes the function marginally less efficient when removing scripts by name or index, but makes our code more maintainable and self-consistent. (The previous version didn't handle strings consistently; it wouldn't try casting them to numbers and using them as indices if they failed to match.)