Editor scripting: Plans for UI operations
Plans for how to implement UI operations in a manner that'll hopefully be pretty efficient; it'll essentially allow some UI operations to be "conditionally asynchronous," with them not blocking the script unless and until the script does something that would be affected by whether or not those operations have completed yet. (For example, the script would be able to asynchronously set font properties on widgets, but if it at some point tried to get a font property from a widget, then it'd have to then wait for those "set" operations to finish.)

Before we implement UI support, we're going to take all of the data that DovahKit is capable of fully loading, and we're going to make it all available to scripts. Doing that now will allow us to identify any pain points in the Lua API backend now, rather than having to come back and deal with it later.

Wrote up some plans for implementing Lua wrappers for extra-data types.