Editor scripting: implemented zombie classes
When a natively-managed object is destroyed, it's not possible for us to track down every reference to it from script and replace these references with nil, nor is it possible to take the script-to-native bindings and make them "impersonate nil" or otherwise test as "falsy." As such, we will instead have to turn the wrappers into "zombies" that error if touched the wrong way, and then implement some sort of way to test whether a wrapper is a zombie.

To that end, I've created a system that, given a wrapper metatable, will iterate over its fields and create a "zombie class." This zombie class has member functions with the same names as those on the original metatable, but calling any will immediately throw an error; similarly, setting any property on an instance of the zombie class will throw an error.

With that implemented, we now need to implement some way for scripts to test whether any given userdata is a zombie.