Editor scripting: zombification fixes; form type enum in script
When we zombify a wrapper, we now clear its form stub and loaded-form pointers. We do this because we're going to be abandoning the wrapper, so we won't be able to clear those pointers later should the form be deleted.

We now offer an enum of form type constants to scripts. Scripts now use signatures rather than raw form type constants; feels more "permanent" that way.