Editor script: work on a form-deletion API
All APIs that can modify form data now run a permissions check. We'll flesh this out later; see the to-do list for details.

Started work on a delete-form API. It's mostly hooked up, except that the S2M message handler doesn't check for it and therefore never actually runs the deletion.