DovahKitCore::delete_form reorganization
Reorganized DovahKitCore::delete_form so that it's possible to run your own code between gathering the forms to delete and actually deleting them. We need this in order to zombify form wrappers in the script VM before actually deleting the underlying forms, so that there aren't extant references to the loaded-form data at the time that the stubs are destroyed.