Editor scripting: form creation
Implemented a Lua API for creating new forms.

Discovered an issue wherein forcibly terminating a script while a message was in transit would result in the sender continuing to run; depending on what the message was meant to do and how it was meant to be used, this could result in it being used with an invalid state, which in the case of (dovah.create_form) led to crashes. This is now fixed: if the script is aborted while a blocking message is in transit, the send-message function now terminates the Lua VM the same way our debug hook would.