Editor scripting: shouts, word of power
Added Lua wrappers for shouts and words of power. This includes being able to access the three words on each shout.

If we wrap a none-stub, game setting form stub, or nullptr form stub, the script VM will no longer pass the wrapper to the script. This makes it easier to return forms to Lua without having to check whether they're nullptr or otherwise something that should be returned as nil, as we'll simply end up returning none (which is functionally the same as nil) instead.

Fixed a crash that could occur when trying to wrap a nullptr form stub. The crash occurred when trying to check the form stub's form type, to know what wrapper class to use.