Form type icon work
Created a form type icon for WOOP (Word Of Power) forms. If I get around to making icons for all the other form types, then we can use them in the Object Window and have our UI be that much more consistent with the Creation Kit.

I went for the obvious joke. I think it's cute!