Plans: temporary/working copies of loaded forms
My attempts at building UI for quests and condition lists have exposed some major shortcomings of the current approach to editing forms. The only viable way to overcome these is to create a mechanism whereby the frontend can create a temporary/working copy of a form's loaded data, make changes to that copy as appropriate, and then "commit" the copy back overtop the original form. I've written down my plans for this task.

This commit also includes tons of incomplete code for a condition list UI. Absolutely none of it is viable.

Once the working-form system is complete, we'll want to revert changes to the code comment for form_reference_t::unmanaged_set, and delete form_reference_t::unmanaged_clear_if.