FormDialogWorkingCopyBase and similar helpers
Created the FormDialogWorkingCopyBase class, and made the Quest dialog a subclass of it.

Added (import) and (commit) functions to FormListListview, to simplify the process of extracting and saving data to forms.

Created the (cobb::qt::bind) and (cobb::qt::unbind) functions, which serve as a way to simplify the task of synchronizing Qt controls to simple values.