Condition UI: rough draft of condition edit window; to-do list
Built a rough draft of our condition-editing window. It's larger than the equivalent window in the Creation Kit, but it's a bit cleaner, and it should be possible to edit most condition arguments without having to open child windows for them.

Did a lot of testing in the Creation Kit and determined: the precise meaning and behavior of conditions' "Use Aliases" and "Use Pack Data" flags; how exactly the Creation Kit exposes aliases and package data when they're set as the "Run On" entity for a condition; what event data a condition can run on.

Checked the CK wiki and saw that "GetIsID" is the default function for new conditions, for convenience. Made a note of this, as it's a useful behavior to mimic.