Started work on the condition editing UI
Changed the sizing of some elements in the Quest dialog, so that the dialogue condition list is a little more comfortable to view.

You can now double-click a condition in a condition list in order to open an edit modal. The edit modal currently displays some values, but cannot save changes to any values: it can display the run-on-type, "swap subject and target" flag, unction, comparison operator, comparison float operand, and "or" flag.