Fixes to ref picker window
The RefPickerButton widget now initializes a spawned window with the pre-existing selection, if any.

Fixed the ref-picker window: it should now have a consistent size even if it already has a value when opened; and when setting a reference that isn't part of the selected cell, the desired reference's parent cell should now properly be selected.