Condition UI: run-on selection
Created a (condition_context) object in the backend, which can be used to quickly and easily retrieve a condition's owning package and quest.

When a condition's "Run On" entity is set to Reference or Alias, you can now select the specific reference or alias. Code exists for properly handling Player as well. Skeletal code is ready for package data when we have that implemented.

Wrote down some plans for a scriptable raster-drawing widget, for when we get to implementing UI scripting.