Support for story manager events in conditions
Created definitions for story manager event data. Moved the existing definitions to a different file and deleted the incomplete and now-empty StoryManagerEventNode.h file.

Made changes to ensure correct endianness on event-related fields in quests.

The condition editor now allows you to select an event member when setting Event Data as the run-on entity.