Fixes to corrupt data in conditions
Fixed two bugs in the condition backend that led to corrupt data in conditions: std::array doesn't default-initialize structs unless you tell it to or it just feels like it that day; and the (condition_arg_value) union default-initialized the wrong member, leading to broken pointers (and MSVC didn't respond to my changing this until I rebuilt the entire damn solution).

Got rid of (story_event_member_code) and friends in an attempt to identify and diagnose the above problem. Turns out, I didn't need to. Will I bring them back? I don't know. So far, they're not needed.

The condition list UI now displays the name of the event data that a condition will run on.