Condition UI work - ran into obstacles
Started work on the code for editing condition parameters, but I've run into several obstacles, which I've added to the to-do list. Ultimately, the code for storing parameters in the UI is going to be pretty darned complicated due to all the edge-cases surrounding parameter types and values; and the current implementation of working copies of forms isn't compatible with things like GetStageDone (or quest aliases and the like in general, actually).