Attempts at handling INFO placement within DIALs
Contains incomplete code for handling the placement of INFOs within their parent DIALs' topic info lists, as well as reverse-engineered findings on how the game does it.

In order to properly handle INFO/PNAM == 0xFFFFFFFF in a partial-flagged INFO override (which should retain the previous position), we'll need to change how we handle overrides. Currently, when we load an override, we delete the overridden stub after copying its file list to the overriding stub, and we store the overriding stub; this means we'll lose the INFO's place within its parent's info list (which is a problem if the parent has not changed). This needs to change.