Fixed crash on loading overrides; fixed handling of INFO/PNAM
Fixed a crash that occurred when loading overrides; it stemmed from us overwriting the existing form stub's flags, most notably the flag indicating that it has multiple source files. As it happens, none of the form stub flags need to be copied from the new stub, so we no longer do that at all.

Modified (file_load_order::accept_form_stub) to no longer try to set the (is_hardcoded) flag on the new form stub. That flag should already be set on hardcoded forms when they're created via (_accept_hardcoded_stub), and since overrides now keep the old stub instead of the new stub, the flag is no longer needed on the new stubs.

Fixed handling of INFO/PNAM when loading an override that moves the info within its parent topic.