TopicInfo now has all needed functions
When the file writer decides to write a form, it is now capable of communicating (via the form-save interface) to that form a pointer to its previous-sibling's stub. TopicInfos need this in order to serialize PNAM properly.

Added (tes_file_writing::record::write_formID_subrecord(uint32_t, const form_stub*, bool)) to aid with serializing PNAM.

Split TopicInfo's link-to and condition lists into "locked" and "normal" variants. The former contain data that cannot be edited because it comes from a partial override.

Wrote clear, clone, save, and sever functions for TopicInfo.