Form stubs now own and manage working copies
Working copies are now managed via form stubs: a form may have at most one working copy, and a pointer to the working copy exists on the form stub. Functions to create and commit working copies have been moved from Form to form_stub, and form_stub has been made a friend of Form in order to be able to access the internals needed for working copy functionality.

Added signals to DovahKitCore for working copy commits and deletions.

Simplified some of the code for DialogWorkingCopyBase, and updated it to match how working copies are now managed and to send the appropriate working copy signals on DovahKitCore.

Fixed TopicInfo::save: it now returns true.