Form type checks on and near form stubs
Added type-checking for dovah::loaded_form_ptr: if a smart pointer is templated on anything other than the base Form type, then it will validate a form stub's form type before accepting the stub. As a result of this change, dovah::loaded_form_ptr::unwrap itself no longer needs any special logic.

Made form_stub::get_working_copy a templated function that defaults to the base Form type, and added form type checking to it.

Added form_stub::get_working_or_stable_copy as a helper function, to save callers some of the burden of doing their own type checks.

The condition-editing UI now prefers quests' working copies, if any, when retrieving a list of their stages for a quest-stage parameter's possible values.

Started work on an internal document meant to describe user-facing features that we offer and the Creation Kit does not. This is mainly intended for incidental features or things that I might forget or be unaware of, such as BSA loading (the CK only loads official BSAs by default, and I only just learned this), but I'll probably add all of our other stuff to it as well.