Fixes and to-do notes for condition parameter widget
The value() function should consistently return a uint8_t QVariant when dealing with character-type condition arguments, even when a textbox is being used to edit them.

We now properly reset the max length on the textbox, so switching from a non-axis character-type parameter to a string-type parameter won't leave the max length stuck at 1.

Added to-do notes to the header file.