Started work on rebuilding the condition backend
Started work on rebuilding the condition backend, to make it possible to make working copies of individual conditions and to make conditions easier and safer to work with. This isn't done, so the program will not build in this state.

Moved condition function and parameter-type definitions to the dovah namespace and to the /dovah/data/ folder.

Conditions now use getter/setter member functions for their function ID and parameters; this allows us to validate what's going in and ensure that no phantom form-to-form references are ever created, among other sanity checks.

Conditions are now aware of their containing form; this should simplify the code for working with them, at the cost of introducing some redundancy (i.e. if a form contains four conditions, then it contains four pointers to itself).